requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
flask==2.2.3
gunicorn==20.1.0
selenium==4.10.0
webdriver-manager==3.8.6
cloudscraper==1.2.71
lxml==4.9.3